+++ Modules/Intent.Modules.Application.DependencyInjection/Templates/DependencyInjection/DependencyInjectionTemplatePartial.cs
@@ -50,7 +50,7 @@ namespace Intent.Modules.Application.DependencyInjection.Templates.DependencyInj
 
             @event.MarkAsHandled();
             _containerRegistrationRequests.Add(@event);
-            
+
             foreach (var templateDependency in @event.TemplateDependencies)
             {
                 var template = GetTemplate<IClassProvider>(templateDependency);
@@ -61,7 +61,7 @@ namespace Intent.Modules.Application.DependencyInjection.Templates.DependencyInj
 
                 AddTemplateDependency(templateDependency);
             }
-            
+
             foreach (var ns in @event.RequiredNamespaces)
             {
                 AddUsing(ns);
@@ -77,7 +77,7 @@ namespace Intent.Modules.Application.DependencyInjection.Templates.DependencyInj
 
             @event.MarkAsHandled();
             _serviceConfigurationRequests.Add(@event);
-            
+
             foreach (var templateDependency in @event.TemplateDependencies)
             {
                 var template = GetTemplate<IClassProvider>(templateDependency);
@@ -88,7 +88,7 @@ namespace Intent.Modules.Application.DependencyInjection.Templates.DependencyInj
 
                 AddTemplateDependency(templateDependency);
             }
-            
+
             foreach (var ns in @event.RequiredNamespaces)
             {
                 AddUsing(ns);